Tiny correction; alignment is always in effect
You may even observe it if the frame and its child are exactly the same size.  You can see that for example in 

```swift
struct ContentView: View {
    var body: some View {
        Text("Hello, World!").border(Color.red)
            .frame(width:50, height: 42.3, alignment: .init(horizontal: .center, vertical: .firstTextBaseline))
            .border(Color.blue)
    }
}
```

But aside from this tiny nit, this is an absolutely *wonderful* article; thank you for writing it!  It captures both the mental model and the reasons we designed it that way perfectly.